requirements.txt
fastapi
uvicorn
yfinance
python-dateutil
